test(mass-tools): rename test
